Change double to BigDecimal

+++ src/main/java/it/marco/lastminute/dto/Book.java
@@ -1,19 +1,21 @@
 package it.marco.lastminute.dto;
 
+import java.math.BigDecimal;
+
 public class Book {
 
 	/*
 	 * VARIABLES
 	 */
 
-	public double amount;
-	public double finalPrice;
+	public BigDecimal amount;
+	public BigDecimal finalPrice;
 
 	/*
 	 * CONSTRUCTORS
 	 */
 
-	public Book(double amount) {
+	public Book(BigDecimal amount) {
 
 		this.amount = amount;
 		this.finalPrice = amount;
